codewave/requirements.txt
amqp==5.0.2
appdirs==1.4.3
asgiref==3.3.1
billiard==3.6.3.0
celery==5.0.2
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
defusedxml==0.6.0
distlib==0.3.0
Django==2.2.3
django-allauth==0.41.0
django-celery-results==2.0.0
djangorestframework==3.12.2
filelock==3.0.12
idna==2.9
importlib-metadata==1.6.0
Jinja2==2.10.3
kombu==5.0.2
MarkupSafe==1.1.1
oauthlib==3.1.0
Pillow==7.1.2
prompt-toolkit==3.0.8
python3-openid==3.1.0
pytz==2020.4
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.4.1
urllib3==1.25.3
vine==5.0.0
virtualenv==20.0.20
wcwidth==0.2.5
zipp==3.1.0
